"This series solve an issue with the load broadcaster that reports negative
 values due to an integer wrap around.  While fixing this issue an additional
 change was made so that the load_map would return doubles and not formatted
 string.  This is a better API, safer and better documented."

+++ service/load_broadcaster.cc
@@ -60,7 +60,7 @@ void load_broadcaster::start_broadcasting() {
                 res += i.second->get_stats().live_disk_space_used;
             }
             return res;
-        }, 0, std::plus<int64_t>()).then([this] (int64_t size) {
+        }, int64_t(0), std::plus<int64_t>()).then([this] (int64_t size) {
             gms::versioned_value::factory value_factory;
             return _gossiper.add_local_application_state(gms::application_state::LOAD,
                 value_factory.load(size)).then([this] {
